Fix #9: properly handle nulls in pre-ban permissions
The reason of the issue was that we were sending to the server exactly
the same permission set we've received from it earlier. This
permission set often (always?) contain null for permissions of the new
users, which means we were first disabling them, and then trying to
"reset" these permissions to null (which, in terms of the Telegram.Bot
API, means do not touch them at all).

This commit changes that behavior: from now on, we'll properly handle
the default permissions set by the chat administrators, and fall back
to just setting the permissions to true if the defaults are unavailable
for some reason.